Added endpoints to calcualte  students GPA and credit hours
I added 2 endpoints, one that does the calcualtion for all students and another that calculates it for a single student incase of a change in a specific student's grade you can update his without doing the calculations for all students all over again. I also fixed some a bug in the students search endpoint.